Rename redux package to state since redux package somehow causes webpack to get confused.
